Solved SQLInjection

+++ Source/functions/deleteautoinstaller.php
@@ -5,7 +5,11 @@ include_once '.././configs/isset_session.php';
 
 $id = $_GET['id'];
 
-$sql = "DELETE FROM saveautoinstaller WHERE id = '$id';";
-mysqli_query($db, $sql);
+if(is_int($id)){
+    $sql = "DELETE FROM saveautoinstaller WHERE id = '$id';";
+    mysqli_query($db, $sql);
+}
 
 header('Location: .././index.php');
+
+?>
